Reference/requirements.txt
# Superannuation Transcripts Demo - Python Dependencies
# =====================================================
# These packages are available in Snowflake's anaconda channel:
# https://repo.anaconda.com/pkgs/snowflake/

# Core Streamlit and data processing
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0

# Snowflake connectivity
snowflake-connector-python>=3.0.0
snowflake-snowpark-python>=1.9.0

# Configuration and utilities
tomli>=2.0.0
typing-extensions>=4.0.0

# Data visualization (for charts and graphs)
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Additional utilities
requests>=2.28.0
urllib3>=1.26.0
python-dateutil>=2.8.0

# For better error handling and logging
logging>=0.4.9.6

# JSON handling (usually built-in, but specified for completeness)
# json - built-in Python module

# Optional: For advanced data processing if needed
# scipy>=1.10.0
# scikit-learn>=1.3.0

# Development and testing (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0

# Note: All packages should be available in Snowflake's anaconda channel
# If deploying to Streamlit in Snowflake, ensure compatibility 